[Feat] Redis 기반 게시글 조회수 flush 구현

+++ src/main/java/com/stocat/amumal/post/dto/GetPostResponse.java
@@ -11,7 +11,7 @@ public record GetPostResponse(
     String nickname,
     String profileImage,
     String createdAt,
-    int viewCount,
+    long viewCount,
     int likeCount,
     int commentCount,
     boolean isLiked,
